+++ frontend/src/ui/components/Employee/Dashboard/RecentAssessment.jsx
@@ -1,16 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import { Button, Title } from "../../../styles/reusableElements.styled";
-import { ReactComponent as ArrowDropdown } from "../assets/dropdown-arrow.svg";
-import { ReactComponent as GridSort } from "../assets/gridsort.svg";
-import { ReactComponent as ListSort } from "../assets/listsort.svg";
-import { ReactComponent as ThreeDots } from "../assets/three-dots.svg";
-import useAuth from "../../../hooks/useAuth";
-import axios from "../../../api/axios";
-import { showErrorToast } from "../../../helpers/helper";
-import { Radar } from "react-chartjs-2";
-import TableComponent from "../molecules/TableComponent";
+import axios from "../../../../api/axios";
+import { showErrorToast } from "../../../../helpers/helper";
+import useAuth from "../../../../hooks/useAuth";
+import { Button, Title } from "../../../../styles/reusableElements.styled";
+import TableComponent from "../../molecules/TableComponent";
+import { ReactComponent as ArrowDropdown } from "../../assets/dropdown-arrow.svg";
+import { ReactComponent as GridSort } from "../../assets/gridsort.svg";
+import { ReactComponent as ListSort } from "../../assets/listsort.svg";
+import { ReactComponent as ThreeDots } from "../../assets/three-dots.svg";
 
 const AssessmentList = [
   {
@@ -68,21 +67,10 @@ const DateDropdownItems = [
   { date: "Last Year" },
   { date: "Custom" },
 ];
-const SKillRating = ({ title }) => {
-  return (
-    <Title
-      as="h5"
-      $size="16px"
-      $weight="normal"
-      $lHeight="22px"
-      $color="#142245"
-    >
-      {title}
-    </Title>
-  );
-};
 
-const Dashboard = () => {
+const RecentAssessment = () => {
+  const { auth } = useAuth();
+
   const [assessmentTypeDropdown, setAssessmentTypeDropdown] =
     React.useState(false);
 
@@ -101,59 +89,9 @@ const Dashboard = () => {
     setSortByDateDropdown(!sortByDateDropdown);
   };
 
-  const { auth } = useAuth();
   const [availableAssessment, setAvailableAssessment] = React.useState(null);
-  const [chartDetails, setChartDetails] = React.useState(null);
-
   const [isAvailableAssessmentLoading, setIsAvailableAssessmentLoading] =
     React.useState(true);
-  const [isChartLoading, setIsChartLoading] = React.useState(true);
-  // const [fetchError, setFetchError] = React.useState("");
-
-  //Get Chart Details
-  React.useEffect(() => {
-    const getChartDetails = async () => {
-      try {
-        // const response = await axios.get(`userscore/employee/${auth.id}`);
-        const response = await axios.get(
-          `userscore/employee/aab92dfa-336d-4d63-8c43-7a9826529988`
-        );
-        setChartDetails(response.data);
-        setIsChartLoading(false);
-      } catch (err) {
-        if (!err?.response) {
-          showErrorToast("No Server Response");
-        } else if (err?.response.data.errorState === true) {
-          showErrorToast(err.response.data.message);
-          // setFetchError(err.response.data.message);
-        }
-      }
-    };
-    getChartDetails();
-  }, [auth, chartDetails]);
-
-  //Chart Data Schema
-  const data = {
-    labels: [0, 0, 0, 0],
-    // labels: chartDetails
-    //   ? JSON.parse(chartDetails?.data[0]?.categories?.split("/").join(""))
-    //   : [0, 0, 0, 0],
-    datasets: [
-      {
-        label: "Dataset",
-        data: chartDetails
-          ? chartDetails?.data[0]?.passed_questions
-          : [0, 0, 0, 0],
-        fill: true,
-        backgroundColor: "rgba(255, 99, 132, 0.2)",
-        borderColor: "rgb(255, 99, 132)",
-        pointBackgroundColor: "rgb(255, 99, 132)",
-        pointBorderColor: "#2667FF",
-        pointHoverBackgroundColor: "#2667FF",
-        pointHoverBorderColor: "rgb(255, 99, 132)",
-      },
-    ],
-  };
 
   //Get Available Assessment Details
   React.useEffect(() => {
@@ -176,47 +114,6 @@ const Dashboard = () => {
 
   return (
     <>
-      <StatsContainer>
-        <SkillSection>
-          <Title
-            as="h5"
-            $size="20px"
-            $weight="100"
-            $lHeight="28px"
-            $color="#142245"
-            style={{ marginBottom: "16px" }}
-          >
-            Your Skill Rating
-          </Title>
-
-          {isChartLoading ? (
-            "Loading"
-          ) : (
-            <SkillRatingSection>
-              {/* <div>
-                {chartDetails
-                  ? JSON.parse(
-                      chartDetails?.data[0]?.categories?.split("/").join("")
-                    ).map((item, key) => <SKillRating key={key} title={item} />)
-                  : ""}
-              </div>
-              <div>
-                {chartDetails
-                  ? JSON.parse(chartDetails?.data[0]?.passed_questions).map(
-                      (item, key) => <SKillRating key={key} title={item} />
-                    )
-                  : ""}
-              </div> */}
-            </SkillRatingSection>
-          )}
-        </SkillSection>
-        <ChartSection>
-          {
-            // isChartLoading ? "Loading" :
-            <Radar data={data} />
-          }
-        </ChartSection>
-      </StatsContainer>
       <AssessmentContainer>
         <AssessmentTitle>
           <Title
@@ -340,64 +237,7 @@ const Dashboard = () => {
   );
 };
 
-export default Dashboard;
-
-const StatsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background: #f8fcfe;
-  border: 1px solid #c7e0f4;
-  border-radius: 12px;
-  padding: 24px 66px;
-  width: 100%;
-  gap: 80px;
-
-  ${({ theme }) => theme.breakpoints.down("lg")} {
-    flex-direction: column;
-    justify-content: center;
-    text-align: center;
-    gap: 30px;
-  }
-`;
-
-const SkillSection = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  border: 1px solid #ebf4f9;
-  border-radius: 12px;
-  background-color: #fff;
-  padding: 49px 32px;
-  gap: 12px;
-  min-height: 336px;
-`;
-
-const SkillRatingSection = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: 50px;
-
-  div {
-    text-align: left;
-  }
-`;
-const ChartSection = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #ebf4f9;
-  width: 80%;
-  border: 1px solid #ebf4f9;
-  border-radius: 12px;
-  padding: 35.35px 40px;
-  min-height: 336px;
-  ${({ theme }) => theme.breakpoints.down("lg")} {
-    width: 100%;
-  }
-`;
+export default RecentAssessment;
 
 const AssessmentContainer = styled.div`
   margin-top: 24px;
