+++ frontend/src/ui/components/Employee/index.js
@@ -1,6 +1,7 @@
-export { default as EmployeeDashboard } from "./Dashboard";
 export { default as AssessmentResult } from "./AssessmentResult/AssessmentResult";
 export { default as AssessmentCompleted } from "./AssessmentList/Completed";
 export { default as AssessmentAvailable } from "./AssessmentList/Available";
+export {default as Stats} from "./Dashboard/Stats"
+export { default as RecentAssessment } from "./Dashboard/RecentAssessment";
 // export { default as AssessmentHeader } from "../../../../miscellaneous/TakeAssessment/AssessmentHeader";
 // export { default as AssessmentResultHeader } from "../../../../miscellaneous/TakeAssessment/AssessmentResultHeader";
