+++ frontend/src/ui/pages/Employee/index.js
@@ -1,2 +1,2 @@
 export { default as EmployeeLogin } from "./Auth/Login";
-export {default as Dashboard} from './EmployeeDashboard'
+export { default as EmployeeDashboard } from "./Dashboard";
